decouple db from relayhmon
